Seperated all front-end things into components.
