--- web/src/App.scss
@@ -1,100 +1,0 @@
-//Custom font
-@import url("https://fonts.googleapis.com/css2?family=Open+Sans&family=Roboto&display=swap");
-.App {
-	margin-top: 2em;
-	text-align: center;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	font-family: "Open Sans", sans-serif;
-}
-
-//Login Main Container
-.login {
-	width: 27em;
-	// height: 40em;
-	// min-height: 44em;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	position: relative;
-	z-index: 99;
-	.container {
-		min-height: 42em;
-		display: flex;
-		justify-content: center;
-		align-items: center;
-		background-color: #fff;
-		box-shadow: 0px 0px 12px 2px rgba(15, 15, 15, 0.2);
-		border-radius: 4px;
-		position: relative;
-		z-index: 99;
-		width: 100%;
-		z-index: 99;
-		&.login-pane {
-			transition: all 450ms ease-in-out;
-			height: 45em;
-		}
-	
-		&.register-pane {
-			transition: all 450ms ease-in-out;
-			height: 50em;
-		}
-	}
-	.right-side {
-		display: flex;
-		flex-direction: column;
-		justify-content: center;
-		height: 90%;
-		background-color: #3498db;
-		width: 100%;
-		position: absolute;
-		right: -34%;
-		border-radius: 6px;
-		z-index: 1;
-		transition: all 400ms ease-in-out;
-		cursor: pointer;
-		box-shadow: 0px 0px 12px 2px rgba(15, 15, 15, 0.281);
-
-		&.right {
-			right: -40%;
-			align-items: flex-end;
-			&:hover {
-				right: -45%;
-			}
-		}
-		&.left {
-			right: 40%;
-			align-items: flex-start;
-			&:hover {
-				right: 45%;
-			}
-		}
-
-		.text {
-			font-size: 21px;
-			font-weight: 500;
-			color: #fff;
-			margin-right: 3em;
-			margin-left: 3em;
-		}
-	}
-}
-
-//Button
-.btn {
-	font-size: 21px;
-	padding: 5px 20px;
-	border: 0;
-	background-color: #3498db;
-	color: #fff;
-	border-radius: 3px;
-	transition: all 250ms ease-in-out;
-	cursor: pointer;
-	&:hover {
-		background-color: #2386c8;
-	}
-	&:focus {
-		outline: none;
-	}
-}
