+++ web/src/components/button.jsx
@@ -1,0 +1,13 @@
+import "../styles/button.scss"
+
+import React, { Component } from 'react'
+
+export class Button extends Component {
+    render() {
+        return (
+            <div>
+                <button className="button">Button</button>
+            </div>
+        )
+    }
+}
