+++ web/src/components/footer.jsx
@@ -1,0 +1,13 @@
+import "../styles/footer.scss"
+
+import React, { Component } from 'react'
+
+export class Footer extends Component {
+    render() {
+        return (
+            <div className="footer">
+                Footer
+            </div>
+        )
+    }
+}
