+++ web/src/components/header.jsx
@@ -1,0 +1,14 @@
+import "../styles/header.scss"
+import "../styles/account.scss"
+
+import React, { Component } from 'react'
+
+export class Header extends Component {
+    render() {
+        return (
+            <div className="header">
+                Login/Register
+            </div>
+        )
+    }
+}
