+++ web/src/components/index.jsx
@@ -1,0 +1,11 @@
+import "../styles/style.scss";
+
+export {Panel} from "./panel"
+export {Type} from "./type"
+export {Header} from "./header"
+export {Media} from "./media"
+export {Input} from "./input"
+export {Footer} from "./footer"
+export {Button} from "./button"
+export {Login} from "./login"
+export {Register} from "./register"
