--- web/src/components/login/index.jsx
@@ -1,4 +1,0 @@
-import "./style.scss";
-
-export { Login } from "./login";
-export { Register } from "./register";
