+++ web/src/styles/account.scss
@@ -1,0 +1,14 @@
+.accounts {
+    min-height: 42em;
+    //display: flex;
+    //justify-content: center;
+    //align-items: center;
+    background-color: #fff;
+    box-shadow: 0px 0px 12px 2px rgba(15, 15, 15, 0.2);
+    border-radius: 4px;
+    position: relative;
+    z-index: 99;
+    //width: 100%;
+    width: 27em;
+    z-index: 99;
+}
