+++ web/src/styles/app.scss
@@ -1,0 +1,21 @@
+//Custom font
+@import url("https://fonts.googleapis.com/css2?family=Open+Sans&family=Roboto&display=swap");
+.app {
+	margin-top: 2em;
+	text-align: center;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	font-family: "Open Sans", sans-serif;
+}
+
+.panel {
+	width: 27em ;
+	height: 45em;
+	// min-height: 44em;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	position: relative;
+
+}
