+++ web/src/styles/button.scss
@@ -1,0 +1,17 @@
+//Button
+.button {
+	font-size: 21px;
+    padding: 5px 20px;
+	border: 0;
+	background-color: #3498db;
+	color: #fff;
+	border-radius: 3px;
+	transition: all 250ms ease-in-out;
+	cursor: pointer;
+	&:hover {
+		background-color: #2386c8;
+	}
+	&:focus {
+		outline: none;
+	}
+}
