+++ web/src/styles/footer.scss
@@ -1,0 +1,3 @@
+.footer {
+    margin: 2em;
+}
