+++ web/src/styles/header.scss
@@ -1,0 +1,3 @@
+.header {
+    margin: 2em;
+}
