+++ web/src/styles/input.scss
@@ -1,0 +1,38 @@
+.input {
+    margin: 2em;
+}
+
+form {
+    margin-top: 2em;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+}
+
+.form-group {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    width: fit-content;
+    label {
+        font-size: 20px;
+    }
+}
+
+input {
+    margin-top: 6px;
+    min-width: 18em;
+    height: 37px;
+    padding: 0px 10px;
+    font-size: 16px;
+    font-family: "Open Sans", sans-serif;
+    background-color: #f3f3f3;
+    border: 0;
+    border-radius: 4px;
+    margin-bottom: 31px;
+    &:focus {
+        outline: none;
+        box-shadow: 0px 0px 12px 0.8px #0e81ce;
+    }
+}
