+++ web/src/styles/login.scss
@@ -1,0 +1,6 @@
+.login {
+    flex: 1;
+    flex-direction: row;
+    transition: all 450ms ease-in-out;
+    height: 45em;
+}
