+++ web/src/styles/media.scss
@@ -1,0 +1,8 @@
+.media {
+    margin: 2em;
+}
+
+img {
+    width: 100%;
+    height: 100%;
+}
