+++ web/src/styles/panel.scss
@@ -1,0 +1,20 @@
+//.login
+.panell {
+    
+    //.container
+    min-height: 42em;
+    background-color: #fff;
+    box-shadow: 0px 0px 12px 2px rgba(15, 15, 15, 0.2);
+    border-radius: 4px;
+
+    &.login {
+        transition: all 450ms ease-in-out;
+        height: 45em;
+    }
+
+    &.register {
+        transition: all 450ms ease-in-out;
+        height: 50em;
+    }
+
+}
