+++ web/src/styles/register.scss
@@ -1,0 +1,4 @@
+.register-pane {
+    transition: all 450ms ease-in-out;
+    height: 50em;
+}
