+++ web/src/styles/type.scss
@@ -1,0 +1,39 @@
+
+.type {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    height: 90%;
+    background-color: #3498db;
+    width: 100%;
+    position: absolute;
+    border-radius: 6px;
+    transition: all 400ms ease-in-out;
+    cursor: pointer;
+    box-shadow: 0px 0px 12px 2px rgba(15, 15, 15, 0.281);
+
+    &.right {
+        right: -40%;
+        align-items: flex-end;
+        &:hover {
+            right: -45%;
+        }
+    }
+    &.left {
+        right: 40%;
+        align-items: flex-start;
+        &:hover {
+            right: 45%;
+        }
+    }
+
+    
+}
+.text {
+    font-size: 21px;
+    font-weight: 500;
+    color: #fff;
+    margin-right: 3em;
+    margin-left: 3em;
+    align-content: right;
+}
